Simplify the behavior of the arithmetic entail utility in strings (#11538)
By default, we no longer use approximation reasoning to recursively
prove that an approximation holds.

The motivation for this is proof holes that are very difficult to reason
about.

The old behavior can be reenabled with expert option
`--strings-rec-arith-approx`.

I tested this change on SMT-LIB and an industrial set with very little
performance differences -5+0 (all 5 benchmarks close to timeout).

Also eliminates another recursive call to the rewriter in the strings
rewriter.

This reduces the set of rewrite steps for strings in our "safe"
configuration from 35 -> 29.