java: Change search order for native libraries (#11560)
After this change, the Java API first attempts to locate native
libraries as resources within a JAR before searching system paths. If
the Java API is packaged in a self-contained JAR, it will prioritize the
native libraries within the JAR, as they are more likely to be
compatible with the API version included in the JAR.

This PR also fixes and enhances error handling when searching for and
loading the native libraries.