Simplifications and improvements for ARITH_MULT_SIGN (#11476)
Makes it consider variables only in order. Furthermore improves the
logic by always inferring the monomial argument is `>` when applicable,
noting that `n != 0 => n*n>0`.

Also makes its argument an AND formula instead of a variable number of
arguments, which is not supported in Eunoia.

This is in preparation for adding its Eunoia definition.