More preparation for eliminating nested calls to rewriter in strings rewriter (#11556)
Behavior (mostly) does not change in this PR, apart from eliminating a
few direct calls to the rewriter.

We no longer pass the rewriter directly to the strings rewriter.

The next step will be to not pass the rewriter to the arith/string
entailment utilities in the main rewriter.

Will performance test this change.