Introduce a primative string utility `Word::hasOverlap` (#11579)
This operation will be at the heart of ~5 forthcoming proof rules. 

This PR deletes the primative `Word::noOverlapWith` (which wasn't
directional and had a more complicated meaning) in favor of this new
utility.

A rewrite that relies on this utility is demoted to the extended
rewriter.

Unit tests are added for this utility.