Add Eunoia definitions of extended RE evaluation (#11581)
This adds the Eunoia definition of the evaluation methods for the
SMT-LIB standard operators `str.replace_re` and `str.replace_re_all`,
and the cvc5-specific extension `str.indexof_re`.

It also updates the strings rewrite to call the reference
implementations of these rules, which was missed in a previous PR.

Note these evaluators call `$str_eval_str_in_re` as a submethod, which
would benefit automatically from future optimizations to this method.