Add CPC and Eunoia definitions for string theory rewrites involving overlap reasoning (#11580)
This adds a four new ad-hoc theory rewrites for strings, including their
internal and Eunoia definitions.

3 of these rules are exercized extensively in SMT-LIB, and all 4 will be
covered in regressions once they are enabled.

They are each based around the notion of computing "overlaps". 

This will be used to elaborate proofs for 4 macro theory rewrites, to be
added in followup PRs.

Note: in their current form, these rules *cannot* be applied to
sequences, as it is non-trivial in general to *prove* the distinctness
of sequence characters.

Furthermore note that the existing definition of
`ProofRule::CONCAT_CPROP` will be later simplified to use the
`$str_has_overlap` utility (it currently computes the *size* of a
maximal overlap using `$str_overlap`, which can be simplified by an
internal elaboration step). The signature definition of this rule at the
moment is theoretically unsound for sequences, since we do not prove the
distinctness of sequence characters (although cvc5 internally will do
so).