More updates to classifications of options (#11576)
Makes the following changes:
- `produceUnsatCores`, `checkUnsatCores` are promoted to "common" as
they are part of our core infrastructure.
- `fullSaturateQuant`/`enumInst` (which are an alias of one another) are
promoted to "common". The justification is that this option is
recommended to be used in addition to any other combination of options
for quantifiers for those that prefer more effort + timeouts to
unknowns.
- Similarly, `decisionMode` is promoted to "common", as it is intended
to be used in combination with all other options, e.g. in portfolios.
- `unconstrainedSimp` and `sortInference` are demoted to "expert" since
they do not meet the requirements of producing *models* (as well as
proofs).
- A new expert option `varEntEqElimQuant` now guards a (seldom used)
rewrite for quantifiers+strings that is hard to justify.