Two fixes for quantifiers prenex proofs (#11487)
The fix involves the use of unused variables to ensure we can miniscope
OR when reversing elaboration of prenex. Both were leading to proof
holes which this PR fixes.

The second ensures we add unique variables for shared subformulas. For
example:
forall x. (forall y. P(y) or forall y. P(y))
we previously would prenex to:
forall x z z. (P(z) or P(z))
now we prenex to
forall x z w. (P(z) or P(w))
While it is technically legal to prenex to the same variable when their
polarity is the same, it is harder to justify and leads to unexpected
shadowing.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>