Refactor proof post processor DSL to avoid subgoal tracking (#11442)
This fixes a bug in how RARE proof reconstruction works for macro
elaboration.

In particular, our subgoal tracking was incorrect if a macro was
elaboarted to a single TRUST step as a subgoal, since the updater
updates the current proof node in-place. The PR
https://github.com/cvc5/cvc5/pull/11441 introduces a case in our
regressions for this.

To fix this, I've refactored the post-processor to a more elegant
version which runs the proof node updater recursively. It tracks the
depth of how many trust steps we've expanded in the current context
using a `d_traversing` vector.

This makes explicit tracking of subgoals unecessary, the RARE proof
reconstruction method is cleaned up as a result.