[proofs] Remove eager check for SAT proof closedness
This test is currently wrong and was leading to issues in a few regressions when
--proof-check=eager is on. It is wrong because the assumptions tracked in
`d_assumptions` are the same assumptions of the SAT proof but only modulo
reordering. This is not currently tracked because for the purpose of building
the proof of the proposition engine the post-processor handles this, since the
CNF proof has the reordering step. To do the closedness check here would require
also tracking this in the SAT proof manager, which I would rather not do.

Removing this check is also not a big issue because we check the closedness of
the proposition proof in the prop proof manager anyway.