Update CPC signature and output to fully support parametric datatypes (#11577)
Takes advantage of recent extension to Ethos on
https://github.com/cvc5/ethos/pull/106.

In particular the proof signature is updated to handle extended
semantics of `eo::dt_constructors` and `eo::dt_selectors`. It is also
updated so that `Nullable` is properly defined as a parameteric
datatype.

We furthermore properly print annotations on *ambiguous* constructor
applications in the proof printer, and the proper printing of Nullable
is now handled.

We can now enable cpc on several regressions.

---------

Co-authored-by: Hans-Jörg <hansjoergschurr@users.noreply.github.com>