Add Eunoia definition for more cases of STRING_REDUCTION (#11587)
This should cover all remaining operators whose reduction could appear
in an SMT-LIB benchmark.

Makes minor changes to the strings utility for the sake of uniformity.

This also inlines the previous definitions of substr/indexof to make the
style of the signature.

Also adds a regression to excersize the strings contain reduction, which
(surprisingly) is not fired in our regressions or on SMT-LIB.