Add Eunoia signature defining translation from sequences to strings (#11612)
The current Eunoia signature for strings is limited in several ways wrt
sequences (6 rules are not applicable, and some give spurious failures).

The side conditions in the new file `SeqToString.eo` will be used to
define a generic evaluator for sequences, as well as fixing the
application of current string rules to sequences.