Fix several Eunoia rules for sequences (#11649)
This makes 6 proof rules in the strings signature applicable to
sequences now.

5 of these rules depended on "overlap" utilities which were not
applicable to sequences; 1 of these rules (concat_cprop) was already
unsoundly assuming sequence characters to be distinct. The overlap
utilities in the strings signature are updated to convert sequences ->
strings if necessary, which should make the rules sound at the
meta-level.

The other rule (multiset containment) is refactored to use the newly
introduced distinct values side condition.