cadical: Make modelValue more robust.
CaDiCaL versions 2.1.0 to 2.1.3 have an issue with CaDiCaL::val(int lit)
where the handling for negative literals is incorrect (already reported to
to the CaDiCaL dev team). In order to avoid this issue and make querying
of values from CaDiCaL more robust on our side wrt to future API changes,
we refactor value querying to query the values of variables only and
we handle its negation depending on the phase of the literal on our side.