Add proof rewrite rule MACRO_BOOL_BV_INVERT_SOLVE (#11687)
This is based on the method for eliminating BV variables from quantified
formulas here:
https://github.com/cvc5/cvc5/blob/main/src/theory/quantifiers/quantifiers_rewriter.cpp#L1204

Also makes the BV inverter independent of options, which is required for
proofs. A seldom used expert option is deleted.

The new method is amenable to be proof elaboration and will be called
instead of the above method in a followup PR.