Fix elaboration of MACRO_SR_PRED_TRANSFORM (#11688)
This PR fixes a (very rare) occurrence of non-closed proofs that occurs
in one benchmark on SMT-LIB and several more when using my dev branch.

The elaboration of MACRO_SR_PRED_TRANSFORM was wrong if:
(1) we are using an extended rewrite method,
(2) the witness form of one of the predicates was itself,
(3) rewriting after the extended rewrite was necessary to show the
predicates to be equal.

In this case, the proof elaboration would neglect to apply the rewriting
again.

The solution is to apply rewriting even if witness form is the identity.
Moreover this change suggests that we should be more fine grained in
which subset of steps (1)-(3) should be applied. This gives initial work
towards this, although this could be further optimized.

A similar issue happens for MACRO_SR_PRED_INTRO.

Unfortunately the benchmark in question takes ~40 seconds in production
and ddsmt does not scale when minimizing it so it is added to regress4.