Fix check for whether a regular expression is constant, guard cases of RegLan in constraints (#11705)
Fixes https://github.com/cvc5/cvc5/issues/11703.

Our check for whether a regular expression was "constant" was incorrect
as it permitted, e.g. APPLY_UF that returns a RegLan.

This issue has been present as early as cvc5 1.0.0. Earlier versions of
cvc5 explicitly forbid constructing constants involving the RegLan type.

This PR furthermore guards cases where terms of return type RegLan are
used in theory combination, which is not supported.

This furthermore makes a fix to the SyGuS solver to not introduce
"evaluation unfolding" lemmas involving RegLan constraints, which now
are guarded.