Add necessary BV rewrites (#11662)
One rewrite taken from PR by @lenianiva.

3 of the other ones are needed for a corner cases of BV rewriting
bvcomp. The other 2 will be necessary to fill proofs for variable
elimination in quantified BV.

Generalizes the bv-mult-slt-mult rules to match the (updated) form of
the rule, following the simplification on
https://github.com/cvc5/cvc5/pull/11704.

Generalizes the RARE concat pull-up rules to match instances of the
AndOrXorConcatPullup rewrite, which further includes 6 more cases.