More updates to safe options (#11664)
`dt-share-sel` is now disabled by default and only enabled when using
SyGuS. It is made an expert option since we do not support proofs for
it. Note this means that --sygus-inst no longer uses shared selectors.
This is necessary since --sygus-inst may be combined with user provided
datatypes.

`cegqi-midpoint` is now enabled by default when safe-options is enabled
and made an expert option. This is because we do not support proofs with
it disabled. Note this leads to slight performance regression on pure
quantified arithmetic logics (LIA and LRA) when safe-mode is enabled.

`cegqi-inf-int` and `cegqi-inf-real` are made expert since we do not
support proofs for them.

The above changes are made so that safe-options with and without proofs
uses the same set of options.

Furthermore: 

`cegqi-bv-ineq` is made expert since it uses invertibility conditions
that are not in the proof signature, nor are found to be useful in
practice.

`ackermann` is made expert since we do not support proofs or models for
it currently, moreover we estimate it is difficult to add proof support
for it from a theory standpoint.