Do not throw an assertion failure when arithmetic model construction fails and a lemma has been sent. (#11565)
We are currently throwing an assertion failure when arithmetic model
construction fails to assert an equality but a lemma (possibly a
buffered NL from the arithmetic solver itself) has already been sent.
This case is benign as we know the theory engine will be called again to
recheck. This PR makes us simply return unsuccessfully in such cases.

The assertion failure is currently firing on
https://github.com/cvc5/cvc5/pull/11453, which makes an unrelated change
to arithmetic.

Note that https://github.com/cvc5/cvc5/pull/11322 recently refactored
the behavior.

On a related note, I have noticed the buffering of the NL solver lemmas
is slightly nonsensical: depending on the order of which is called first
(collectModelInfo or needsLastCallEffort), the arithmetic solver behaves
differently. If the former is called first, we make no attempt to
construct the model whereas if the latter is called first, we do. Making
the behavior consistent (in either direction) leads to multiple
regressions timing out. I will investigate this after this PR is merged.