cmake: Add JNI_HOME variable to specify the path to JNI libraries (#11733)
When cross-compiling, we need both a host Java executable to generate
the cvc5 JNI headers and the JNI Java libraries for the target platform,
which are usually located in different directories. However, both
`FindJava` and `FindJNI` use `JAVA_HOME` to explicitly specify a Java
installation prefix. This PR adds a `JNI_HOME` variable, allowing the
user to specify a separate path for JNI libraries distinct from the Java
executable.

Fixes #11726