Miscellaneous changes to string rewriter (#11779)
This fixes a missing case of rewriting RE inter/union.

It also ensures we do not pass the rewriter to the arithmetic utility,
to avoid rewrites that cannot be constructed. It refactors the
arithmetic entailment utility to do slightly more reasoning that is
independent of the rewriter.

We have no more strings rewrite holes on main after this PR.