Infrastructure and proof rules involving witness terms (#11536)
This is work towards https://github.com/cvc5/cvc5/pull/11461.

This PR sets up the "valid witness proof generator" as means of
constructing witness terms whose proofs can be reconstructed.

The main idea is to carry an annotation (Rule, list of arguments) as an
attribute on the witness term.

Specific implementations of this will be added in forthcoming PRs.