Refactor and minor improvements to MBQI (#11722)
This is in preparation for new techniques in `mbqi-enum`.

This PR:
- Makes it so that we do not fail as strictly when attempting to
construct an instantiation from model values that go beyond the current
domain of uninterpreted sorts. This was often leading to trivial
"unknowns".
- Makes submodules (mbqi-enum) responsible for sending their own
instantiations, which will allow a better policy for dealing with
duplicates in a followup PR.
- Introduces an inference id for mbqi-enum for stats/debugging.
- Limits the instantiation rounds for nested quantified formulas to 5,
to avoid timeouts in MBQI subcalls with nested quantifiers.