Remove 100 unecessary RARE rules (#11792)
This removes 100 RARE rules which do not appear in regressions, SMT-LIB
or in industrial applications. Each of these rules moreover is subsumed
by existing rules, meaning that it is not possible to use these rules
currently.

We require deleting these rules for two reasons:
(1) To simplify the signature
(2) To improve performance, as it reduces the space which the RARE
reconstruction searches. Note that we have some proof holes since RARE
reaches its resource limit unless these rules are deleted.

There are still roughly 20 rules that are unused due to e.g. the
corresponding rewrite being disabled. These will be addressed in future
PRs.

Details on how the rules are subsumed/unecessary are given below:

; Subsumed by ARITH_POLY_NORM:
arith-to-real-elim
bv-neg-mult
bv-neg-add
bv-mult-distrib-const-neg
bv-mult-distrib-const-add
bv-mult-distrib-const-sub
bv-mult-distrib-1
bv-mult-distrib-2
bv-commutative-mul
bv-mul-one
bv-add-zero
bv-add-two
bv-mul-flatten
bv-neg-sub
bv-neg-idemp

; Handled by ARITH_REDUCTION:
arith-div-total-real
arith-div-total-int

; Subsumed by ARITH_POLY_NORM_REL:
arith-geq-norm2

; Subsumed by ARITH_POLY_NORM_REL+EVALUATE:
arith-refl-leq
arith-refl-lt
arith-refl-geq
arith-refl-gt

; Subsumed by ACI_NORM:
arith-mult-flatten
bv-commutative-and
bv-commutative-or
bv-or-zero
bv-or-flatten
bv-and-flatten
bv-bitwise-idemp-1
bv-bitwise-idemp-2
bv-and-one
bv-xor-zero
str-concat-flatten
str-concat-clash-char
str-concat-clash-char-rev
str-concat-flatten-eq
str-concat-flatten-eq-rev
re-concat-emp
re-concat-flatten
re-union-none
re-union-flatten
re-union-dup
re-inter-flatten
re-inter-dup

; Subsumed by ARITH_REDUCTION:
arith-abs-elim-int
arith-abs-elim-real

; Subsumed by ABSORB:
bv-mul-zero
bv-and-zero
bv-or-one
re-concat-none
re-inter-none
bool-or-true
bool-and-false

; subsumed by bv-and-simplify-1,2, bv-or-simplify-1,2:
bv-bitwise-not-and
bv-bitwise-not-or

; Subsumed by bv-ule-eliminate, bv-sle-eliminate:
bv-not-ule
bv-not-sle

; zero_extend 0 is always eliminated:
bv-merge-sign-extend-3

; Subsumed as part of MACRO_ARITH_STRING_PRED_ENTAIL:
str-len-substr-ub1
str-len-substr-ub2

; Subsumed by str-contains-leq-len-eq + str-eq-len-false:
str-contains-lt-len

; Subsumed by str-contains-leq-len-eq:
str-contains-is-emp

; Subsumed by str-replace-find-first-concat:
str-replace-contains-pre

; Subsumed by str-indexof-contains-pre:
str-indexof-find

; Subsumed by str-substr-concat2:
str-substr-len-skip

; Subsumed by re-in-cstring + MACRO_STR_EQ_LEN_UNIFY_PREFIX:
str-in-re-concat-emp

; Subsumed by STR_IN_RE_CONSUME:
str-in-re-strip-prefix
str-in-re-strip-prefix-neg
str-in-re-strip-prefix-sr-single
str-in-re-strip-prefix-sr-single-neg
str-in-re-strip-prefix-srs-single
str-in-re-strip-prefix-srs-single-neg
str-in-re-strip-prefix-s-single
str-in-re-strip-prefix-s-single-neg
str-in-re-strip-prefix-base
str-in-re-strip-prefix-base-neg
str-in-re-strip-prefix-base-s-single
str-in-re-strip-prefix-base-s-single-neg
str-in-re-strip-char
str-in-re-strip-char-s-single
str-in-re-strip-prefix-rev
str-in-re-strip-prefix-neg-rev
str-in-re-strip-prefix-sr-single-rev
str-in-re-strip-prefix-sr-single-neg-rev
str-in-re-strip-prefix-srs-single-rev
str-in-re-strip-prefix-srs-single-neg-rev
str-in-re-strip-prefix-s-single-rev
str-in-re-strip-prefix-s-single-neg-rev
str-in-re-strip-prefix-base-rev
str-in-re-strip-prefix-base-neg-rev
str-in-re-strip-prefix-base-s-single-rev
str-in-re-strip-prefix-base-s-single-neg-rev
str-in-re-strip-char-rev
str-in-re-strip-char-s-single-rev
str-in-re-req-unfold
str-in-re-req-unfold-rev
str-in-re-skip-unfold
str-in-re-skip-unfold-rev
str-in-re-test-unfold
str-in-re-test-unfold-rev