Export ParserEndOfFileException class (#11836)
Fixes #10414 and duplicate #11146.

The function `InteractiveShell::readAndExecCommands()` is designed to
catch `ParserEndOfFile` exceptions thrown by `Lexer::parseError`.
However, if the exception class is not exported and the cvc5 binary is
linked against a shared version of the cvc5 parser library, the
exception is interpreted as its exported superclass, `ParserException`.
This results in unintended behavior.