Optimize arithmetic proof generation with pred transform utility (#11819)
Many of the low-level proof steps in arithmetic are suboptimal in terms
of performance, for two reasons:
(1) They introduce spurious MACRO_SR_PRED_TRANSFORM steps (proving F
from F).
(2) They do not provide an expected result, meaning that proof checking
is invoked for these steps.

This addresses both issues.

This is work towards optimizing the performance of proofs in the
arithmetic theory.