Add macro expansion for MACRO_QUANT_VAR_ELIM_INEQ (#11826)
This expands the macro `MACRO_QUANT_VAR_ELIM_INEQ` which for example
proves:
`(forall ((x Int) (y Int)) (=> (> x a) (P y)) = (forall ((y Int)) (P
y))`

The proof expansion is based on miniscoping `y`, instantiating `x` with
a "max term" e.g. `a+1` in the above example, and rewriting the result.

This avoids having to introduce an new rule for this case.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>