Fix model construction for parameteric datatypes + u sorts + FMF (#11875)
Fixes https://github.com/cvc5/cvc5-projects/issues/622.

This fixes an issue with parameteric datatypes which was responsible for
the issue, namely that the code for "ordering" types based on their
dependencies was incorrect for parameteric datatypes. In particular, the
bug was caused by enumeration a parameteric datatype over an
uninterpreted sort (after instantiating parameters) being enumerated
before the uninterpreted sort.

FYI @barrettcw