Improve model construction for uninterpreted sorts + FMF (#11878)
We previously relied on a hack in the core model class that allowed
uninterpreted sorts to be freely enumerated, despite being "interpreted
finite" when finite model finding is on.

This was wrong since we may run out of values e.g. when datatypes over
uninterpreted sorts are present.

Since treating them as finite is also wrong (since then all terms would
be assigned the same value), we assign them to unique values at the
*beginning* of model construction.

FYI, another solution would be to make the UF solver responsible for
assigning distinct values to initial equivalence classes. However, this
breaks many invariants, as it introduces "uninterpreted sort values" in
the equality engine of the model, which incorrectly escape to parts of
the FMF algorithm. Instead this PR silently assigns these values in the
constant rep map of the model.