Update set defaults to ensure proofs do not impact safe mode builds (#11895)
This makes one minor change to safe-mode: bitblast internal should be
the default in `safe-mode=safe`, as it is required when proofs are
enabled. (The strategy is not changed when `safe-mode=stable`).

Moreover, we want to ensure that enabling/disabling proofs has minimal
impact, i.e. no changes to solving strategies.

This can be modified once the default bitblasting BV solver has proofs.